refactor : 컴포넌트 위치 변경, 절대경로 설정 #22
