+++ client/tsconfig.json
@@ -21,7 +21,8 @@
       "@/stories/*": ["stories/*"],
       "@/constants/*": ["constants/*"],
       "@/types/*": ["types/*"],
-      "@/utils/*": ["utils/*"]
+      "@/utils/*": ["utils/*"],
+      "@/components/*": ["components/*"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
